requirements.txt
pystray>=0.19
Pillow>=10.0
pywin32>=306
pyinstaller>=6.0
